Largest prime factor of a given number
A prime number has only two factors: 1 and the number itself.
It is important to note however that, a prime number can have four factors in the real number space viz., 1, -1, the number and negative of the number. I haven't considered the same in the above algorithm and have resorted to 2 factors.